polishing: rename files with wrong case, and trim some trailing spaces
