+++ src/storages/inMemory/__tests__/UniqueKeysCacheInMemory.spec.ts
@@ -1,9 +1,9 @@
 // @ts-nocheck
-import { UniqueKeysCacheInMemory } from '../uniqueKeysCacheInMemory';
+import { UniqueKeysCacheInMemory } from '../UniqueKeysCacheInMemory';
 
 test('UNIQUE KEYS CACHE IN MEMORY / should incrementally store values, clear the queue, and tell if it is empty', () => {
   const c = new UniqueKeysCacheInMemory();
-  
+
   // queue is initially empty
   expect(c.pop()).toEqual({keys:[]});
   expect(c.isEmpty()).toBe(true);
