+++ src/storages/inRedis/__tests__/UniqueKeysCacheInRedis.spec.ts
@@ -1,18 +1,18 @@
 //@ts-nocheck
 
 import Redis from 'ioredis';
-import { UniqueKeysCacheInRedis } from '../uniqueKeysCacheInRedis';
+import { UniqueKeysCacheInRedis } from '../UniqueKeysCacheInRedis';
 import { loggerMock } from '../../../logger/__tests__/sdkLogger.mock';
 import { RedisMock } from '../../../utils/redis/RedisMock';
 
 describe('UNIQUE KEYS CACHE IN REDIS', () => {
-  
+
   test('UNIQUE KEYS CACHE IN REDIS / should incrementally store values, clear the queue, and tell if it is empty', async () => {
     const connection = new Redis();
     const key = 'unique_key_post';
-    
+
     const cache = new UniqueKeysCacheInRedis(loggerMock, key, connection);
-    
+
     // queue is initially empty
     expect(cache.pop()).toEqual({keys:[]});
     expect(cache.isEmpty()).toBe(true);
@@ -45,7 +45,7 @@ describe('UNIQUE KEYS CACHE IN REDIS', () => {
     cache.clear();
     expect(cache.pop()).toEqual({ keys: [] });
     expect(cache.isEmpty()).toBe(true);
-    
+
     await connection.quit();
   });
 
@@ -53,7 +53,7 @@ describe('UNIQUE KEYS CACHE IN REDIS', () => {
     let cbCalled = 0;
     const connection = new Redis();
     const key = 'unique_key_post';
-    
+
     const cache = new UniqueKeysCacheInRedis(loggerMock, key, connection, 3); // small uniqueKeysCache size to be reached
     cache.setOnFullQueueCb(() => { cbCalled++; cache.clear(); });
 
@@ -84,58 +84,58 @@ describe('UNIQUE KEYS CACHE IN REDIS', () => {
     const key = 'unique_key_post';
     // Clean up in case there are still keys there.
     await connection.del(key);
-    
-    const cache = new UniqueKeysCacheInRedis(loggerMock, key, connection, 20);  
+
+    const cache = new UniqueKeysCacheInRedis(loggerMock, key, connection, 20);
     cache.track('key1', 'feature1');
     cache.track('key2', 'feature2');
     cache.track('key1', 'feature3');
     cache.track('key2', 'feature3');
-    
+
     await cache.postUniqueKeysInRedis();
-    
+
     connection.lrange(key, 0, 10, async (err, data) => {
       const expected = [
         JSON.stringify({'f': 'feature1', 'ks': ['key1']}),
         JSON.stringify({'f': 'feature2', 'ks': ['key2']}),
         JSON.stringify({'f': 'feature3', 'ks': ['key1', 'key2']})
       ];
-      
+
       expect(data).toStrictEqual(expected);
-      
+
       await connection.del(key);
       await connection.quit();
     });
-    
-    
+
+
   });
-  
+
   test('UNIQUE KEYS CACHE IN REDIS / start and stop task', (done) => {
     const connection = new RedisMock();
     const key = 'unique_key_post';
     const refreshRate = 200;
-    
-    const cache = new UniqueKeysCacheInRedis(loggerMock, key, connection , undefined, refreshRate);  
+
+    const cache = new UniqueKeysCacheInRedis(loggerMock, key, connection , undefined, refreshRate);
     cache.track('key1', 'feature1');
     cache.track('key2', 'feature2');
     cache.track('key1', 'feature3');
     cache.track('key2', 'feature3');
-    
+
     expect(connection.pipeline).not.toBeCalled();
-    
+
     cache.start();
-    
+
     setTimeout(() => {
       expect(connection.pipeline).toBeCalledTimes(1);
       cache.stop();
       expect(connection.pipeline).toBeCalledTimes(2);
       cache.track('key3', 'feature4');
     }, refreshRate + 30);
-    
+
     setTimeout(() => {
-      
+
       expect(connection.pipeline).toBeCalledTimes(2);
       cache.start();
-      
+
       setTimeout(() => {
         expect(connection.pipeline).toBeCalledTimes(3);
         cache.stop();
@@ -143,16 +143,16 @@ describe('UNIQUE KEYS CACHE IN REDIS', () => {
       }, refreshRate + 30);
     }, 2 * refreshRate + 30);
   });
-  
+
   test('UNIQUE KEYS CACHE IN REDIS / Should call "onFullQueueCb" when the queue is full.', async () => {
     const connection = new Redis();
     const key = 'unique_key_post';
-    
+
     // Clean up in case there are still keys there.
     await connection.del(key);
-    
-    const cache = new UniqueKeysCacheInRedis(loggerMock, key, connection, 3);  
-    
+
+    const cache = new UniqueKeysCacheInRedis(loggerMock, key, connection, 3);
+
     cache.track('key1', 'feature1');
     cache.track('key1', 'feature1');
     cache.track('key1', 'feature1');
@@ -161,34 +161,34 @@ describe('UNIQUE KEYS CACHE IN REDIS', () => {
     });
     cache.track('key1', 'feature2');
     cache.track('key2', 'feature3');
-    
+
     const expected1 = [
       JSON.stringify({'f': 'feature1', 'ks': ['key1']}),
       JSON.stringify({'f': 'feature2', 'ks': ['key1']}),
       JSON.stringify({'f': 'feature3', 'ks': ['key2']})
     ];
-    
+
     connection.lrange(key, 0, 10, (err, data) => {
       expect(data).toStrictEqual(expected1);
     });
-    
+
     const expected2 = [
       ... expected1,
       JSON.stringify({'f': 'feature4', 'ks': ['key2']}),
       JSON.stringify({'f': 'feature5', 'ks': ['key3']}),
       JSON.stringify({'f': 'feature6', 'ks': ['key2']})
     ];
-    
+
     cache.track('key2', 'feature4');
     cache.track('key2', 'feature4');
     cache.track('key3', 'feature5');
     cache.track('key2', 'feature6');
-    
+
     connection.lrange(key, 0, 10, async (err, data) => {
       expect(data).toStrictEqual(expected2);
       await connection.del(key);
       await connection.quit();
     });
-    
+
   });
 });
