+++ src/sync/submitters/__tests__/uniqueKeysSubmitter.spec.ts
@@ -1,7 +1,7 @@
 import { uniqueKeysSubmitterFactory } from '../uniqueKeysSubmitter';
 import { loggerMock } from '../../../logger/__tests__/sdkLogger.mock';
-import { UniqueKeysCacheInMemory } from '../../../storages/inMemory/uniqueKeysCacheInMemory';
-import { UniqueKeysCacheInMemoryCS } from '../../../storages/inMemory/uniqueKeysCacheInMemoryCS';
+import { UniqueKeysCacheInMemory } from '../../../storages/inMemory/UniqueKeysCacheInMemory';
+import { UniqueKeysCacheInMemoryCS } from '../../../storages/inMemory/UniqueKeysCacheInMemoryCS';
 
 const imp1 = {
   feature: 'someFeature',
@@ -19,11 +19,11 @@ function getParams(uniqueKeysCache: any) {
   const params = {
     settings: { log: loggerMock, scheduler: { uniqueKeysRefreshRate: 200 }, core: { key: undefined } },
     storage: { uniqueKeys: uniqueKeysCache },
-    splitApi: { 
+    splitApi: {
       postUniqueKeysBulkCs: jest.fn(() => Promise.resolve()),
       postUniqueKeysBulkSs: jest.fn(() => Promise.resolve()) }
   };
-  
+
   return params;
 }
 
@@ -54,7 +54,7 @@ describe('uniqueKeys submitter', () => {
       done();
     }, params.settings.scheduler.uniqueKeysRefreshRate + 10);
   });
-  
+
   test('doesn\'t drop items from cache when POST is resolved CS', (done) => {
     const uniqueKeysCache = new UniqueKeysCacheInMemoryCS(4);
     const params: any = getParams(uniqueKeysCache);
@@ -81,5 +81,5 @@ describe('uniqueKeys submitter', () => {
       done();
     }, params.settings.scheduler.uniqueKeysRefreshRate + 10);
   });
-  
+
 });
